requirements.txt
dask-expr==1.1.7
fastapi==0.111.1
httptools==0.6.1
linkify-it-py==2.0.3
mdit-py-plugins==0.4.1
ollama==0.2.1
openpyxl==3.1.5
pipdeptree==2.23.1
PyMuPDF==1.24.7
watchfiles==0.22.0
websockets==12.0
weaviate-client==4.6.7
langchain==0.2.10
langgraph==0.1.9